requirements.txt
plaid-python
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
python-dotenv
requests
pytest
